ocr/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
ci-info==0.2.0
click==8.0.1
distro==1.6.0
etelemetry==0.2.2
filelock==3.0.12
google-api-core==2.0.1
google-auth==2.0.2
google-cloud==0.34.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
isodate==0.6.0
lxml==4.6.3
networkx==2.6.2
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.2
pdf2image==1.16.0
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
rdflib==6.0.0
requests==2.26.0
rsa==4.7.2
scipy==1.7.1
simplejson==3.17.5
six==1.16.0
tabula-py==2.3.0
traits==6.2.0
urllib3==1.26.6
